fix overflow hiding distant sprites (#1273)
https://github.com/kraflab/dsda-doom/commit/de7bf11a63200e3667353157eaa26ca10625898a

+++ src/r_things.c
@@ -513,7 +513,7 @@ void R_ProjectSprite (mobj_t* thing)
   tx = -(gyt+gxt);
 
   // too far off the side?
-  if (abs(tx)>(tz<<2))
+  if (abs(tx)>((int64_t) tz<<2))
     return;
 
     // decide which patch to use for sprite relative to player
